requirements.txt
h5py==2.8.0
pyzmq==17.1.2
numpy==1.16.4
flatbuffers==1.11
ont-fast5-api==3.0.1
